Refactoring internal/service package
