+++ internal/v1/formula/mock_test.go
@@ -3,9 +3,7 @@ package formula
 import (
 	"context"
 
-	"github.com/moorara/cherry/internal/service/git"
-	"github.com/moorara/cherry/internal/service/github"
-	"github.com/moorara/cherry/internal/service/semver"
+	"github.com/moorara/cherry/internal/service"
 )
 
 type mockUI struct {
@@ -56,13 +54,13 @@ type mockGit struct {
 	IsCleanOutResult bool
 	IsCleanOutError  error
 
-	GetRepoOutRepo  *git.Repo
+	GetRepoOutRepo  *service.Repo
 	GetRepoOutError error
 
-	GetBranchOutBranch *git.Branch
+	GetBranchOutBranch *service.Branch
 	GetBranchOutError  error
 
-	GetHEADOutCommit *git.Commit
+	GetHEADOutCommit *service.Commit
 	GetHEADOutError  error
 
 	CommitInMessage string
@@ -80,15 +78,15 @@ func (m *mockGit) IsClean() (bool, error) {
 	return m.IsCleanOutResult, m.IsCleanOutError
 }
 
-func (m *mockGit) GetRepo() (*git.Repo, error) {
+func (m *mockGit) GetRepo() (*service.Repo, error) {
 	return m.GetRepoOutRepo, m.GetRepoOutError
 }
 
-func (m *mockGit) GetBranch() (*git.Branch, error) {
+func (m *mockGit) GetBranch() (*service.Branch, error) {
 	return m.GetBranchOutBranch, m.GetBranchOutError
 }
 
-func (m *mockGit) GetHEAD() (*git.Commit, error) {
+func (m *mockGit) GetHEAD() (*service.Commit, error) {
 	return m.GetHEADOutCommit, m.GetHEADOutError
 }
 
@@ -118,22 +116,22 @@ type mockGithub struct {
 	CreateReleaseInCtx         context.Context
 	CreateReleaseInRepo        string
 	CreateReleaseInBranch      string
-	CreateReleaseInVersion     semver.SemVer
+	CreateReleaseInVersion     service.SemVer
 	CreateReleaseInDescription string
 	CreateReleaseInDraf        bool
 	CreateReleaseInPrerelease  bool
-	CreateReleaseOutRelease    *github.Release
+	CreateReleaseOutRelease    *service.Release
 	CreateReleaseOutError      error
 
 	GetReleaseInCtx      context.Context
 	GetReleaseInRepo     string
-	GetReleaseInVersion  semver.SemVer
-	GetReleaseOutRelease *github.Release
+	GetReleaseInVersion  service.SemVer
+	GetReleaseOutRelease *service.Release
 	GetReleaseOutError   error
 
 	UploadAssetsInCtx     context.Context
 	UploadAssetsInRepo    string
-	UploadAssetsInVersion semver.SemVer
+	UploadAssetsInVersion service.SemVer
 	UploadAssetsInAssets  []string
 	UploadAssetsOutError  error
 }
@@ -146,7 +144,7 @@ func (m *mockGithub) BranchProtectionForAdmin(ctx context.Context, repo, branch 
 	return m.BranchProtectionForAdminOutError
 }
 
-func (m *mockGithub) CreateRelease(ctx context.Context, repo, branch string, version semver.SemVer, description string, draf, prerelease bool) (*github.Release, error) {
+func (m *mockGithub) CreateRelease(ctx context.Context, repo, branch string, version service.SemVer, description string, draf, prerelease bool) (*service.Release, error) {
 	m.CreateReleaseInCtx = ctx
 	m.CreateReleaseInRepo = repo
 	m.CreateReleaseInBranch = branch
@@ -157,14 +155,14 @@ func (m *mockGithub) CreateRelease(ctx context.Context, repo, branch string, ver
 	return m.CreateReleaseOutRelease, m.CreateReleaseOutError
 }
 
-func (m *mockGithub) GetRelease(ctx context.Context, repo string, version semver.SemVer) (*github.Release, error) {
+func (m *mockGithub) GetRelease(ctx context.Context, repo string, version service.SemVer) (*service.Release, error) {
 	m.GetReleaseInCtx = ctx
 	m.GetReleaseInRepo = repo
 	m.GetReleaseInVersion = version
 	return m.GetReleaseOutRelease, m.GetReleaseOutError
 }
 
-func (m *mockGithub) UploadAssets(ctx context.Context, repo string, version semver.SemVer, assets []string) error {
+func (m *mockGithub) UploadAssets(ctx context.Context, repo string, version service.SemVer, assets []string) error {
 	m.UploadAssetsInCtx = ctx
 	m.UploadAssetsInRepo = repo
 	m.UploadAssetsInVersion = version
@@ -191,19 +189,19 @@ func (m *mockChangelog) Generate(ctx context.Context, gitTag string) (string, er
 	return m.GenerateOutResult, m.GenerateOutError
 }
 
-type mockSemVerManager struct {
-	ReadOutSemVer semver.SemVer
+type mockVersionManager struct {
+	ReadOutSemVer service.SemVer
 	ReadOutError  error
 
 	UpdateInVersion string
 	UpdateOutError  error
 }
 
-func (m *mockSemVerManager) Read() (semver.SemVer, error) {
+func (m *mockVersionManager) Read() (service.SemVer, error) {
 	return m.ReadOutSemVer, m.ReadOutError
 }
 
-func (m *mockSemVerManager) Update(version string) error {
+func (m *mockVersionManager) Update(version string) error {
 	m.UpdateInVersion = version
 	return m.UpdateOutError
 }
